LayoutTests/imported/w3c/web-platform-tests/service-workers/cache-storage/serviceworker/cache-storage.https-expected.txt

CONSOLE MESSAGE: line 153: TypeError: undefined is not an object (evaluating 'registration.unregister')

FAIL CacheStorage assert_unreached: unregister should not fail: serviceWorker.getRegistration() is not yet implemented Reached unreachable code
